+++ doc/changelog.md
@@ -10,7 +10,7 @@
   - Add support for setting a custom maintainer in images (#117)
   - Allow containers to be named using `<namingStrategy>alias</namingStrategy>` when started (#48)
   - Add new global property 'docker.verbose' for switching verbose image build output (#36)
-  - Ass support for environment variables specified in a property file (#128)
+  - Add support for environment variables specified in a property file (#128)
   - Documentation improvements (#107, #121)
   - Allow to use a dockerFileDir without any assembly
   - Switched off strict host address checking in certificates when SSL is used