+++ doc/changelog.md
@@ -10,7 +10,8 @@
   - Added watch mode for images with `-Ddocker.watch` (#141)
   - Added support for inline assemblies (#157, #158)
   - Add support for variable substitution is environment declarations (#137)
-  
+  - Use Tar archive as intermediate container when creating image (#139)  
+   
 * **0.11.3**
   - Add support for removeVolumes in `docker:stop` configuration (#120)
   - Add support for setting a custom maintainer in images (#117)